feat: modernize CLI across all 5 phases (#103)
Phase 1 - Foundation fixes:
- Add rich>=13.0.0 and questionary>=2.0.0 to dependencies
- Replace ANSI codes in ui_components.py with Rich Console
- Context-aware config discovery (walks parent dirs for assignment.conf)
- Remove unused AssignmentSetup import from cli.py

Phase 2 - Interactive prompts:
- Add classdock/utils/prompt.py (questionary wrapper with TTY detection)
- Replace numbered input() selector with questionary arrow-key menu
- Update InputHandler to delegate to prompt.py in TTY environments
- Replace typer.confirm() calls with prompt_confirm() + non-TTY fallback
- Add classdock/interactive.py for no-args interactive mode
- Change no_args_is_help=False; use ctx.invoked_subcommand for detection

Phase 3 - Shortcuts and dashboard:
- Add top-level shortcuts: run, setup, fetch, status, token, completion
- Add classdock/dashboard.py with Rich status panel
- Set rich_markup_mode="rich" on app with visual command tree in docstring

Phase 4 - Polish:
- Add classdock/first_run.py for guided first-run wizard
- Add classdock/utils/error_display.py with Rich error panels

Phase 5 - Structural refactor:
- Split 2600-line cli.py into per-group modules under classdock/commands/
- cli.py reduced from 2712 to 477 lines (app setup + shortcuts only)
- New modules: assignments.py, repos.py, secrets.py, automation.py,
  config.py, roster.py with shared helpers in _helpers.py
- No behavior changes; all 830 tests pass

Also fixes:
- roster sync "no such table" bug: always initialize DB schema
- Update test patch targets to new module locations

Closes #103
