+++ classdock/commands/__init__.py
@@ -1,0 +1,6 @@
+"""
+Command group modules for ClassDock CLI.
+
+Each module defines a Typer sub-app and all its commands.
+cli.py imports the sub-apps and registers them on the root app.
+"""
