+++ pyproject.toml
@@ -36,6 +36,8 @@ dependencies = [
     "PyGithub>=1.59.0",
     "cryptography>=41.0.0",
     "PyNaCl>=1.5.0",
+    "rich>=13.0.0",
+    "questionary>=2.0.0",
 ]
 
 [project.urls]
@@ -75,6 +77,8 @@ include = [
 
 [tool.poetry.dependencies]
 python = "^3.10"
+rich = ">=13.0.0"
+questionary = ">=2.0.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
