+++ tests/test_cli_secrets_force.py
@@ -49,7 +49,7 @@ def mock_global_config():
 class TestSecretsAddCLIForceFlag:
     """Test the --force flag for secrets add CLI command."""
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     @patch('classdock.services.secrets_service.get_global_config')
     @patch('classdock.secrets.github_secrets.GitHubSecretsManager')
     def test_secrets_add_without_force_flag(self, mock_manager_class, mock_get_config_service, mock_get_config_cli, runner, mock_global_config):
@@ -75,7 +75,7 @@ class TestSecretsAddCLIForceFlag:
         call_kwargs = mock_manager.add_secrets_from_global_config.call_args.kwargs
         assert call_kwargs.get('force_update', False) is False
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     @patch('classdock.services.secrets_service.get_global_config')
     @patch('classdock.secrets.github_secrets.GitHubSecretsManager')
     def test_secrets_add_with_force_flag(self, mock_manager_class, mock_get_config_service, mock_get_config_cli, runner, mock_global_config):
@@ -102,7 +102,7 @@ class TestSecretsAddCLIForceFlag:
         call_kwargs = mock_manager.add_secrets_from_global_config.call_args.kwargs
         assert call_kwargs.get('force_update') is True
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     @patch('classdock.services.secrets_service.get_global_config')
     @patch('classdock.secrets.github_secrets.GitHubSecretsManager')
     def test_secrets_add_with_force_short_flag(self, mock_manager_class, mock_get_config_service, mock_get_config_cli, runner, mock_global_config):
@@ -128,7 +128,7 @@ class TestSecretsAddCLIForceFlag:
         call_kwargs = mock_manager.add_secrets_from_global_config.call_args.kwargs
         assert call_kwargs.get('force_update') is True
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     @patch('classdock.services.secrets_service.get_global_config')
     @patch('classdock.secrets.github_secrets.GitHubSecretsManager')
     def test_secrets_add_force_with_multiple_repos(self, mock_manager_class, mock_get_config_service, mock_get_config_cli, runner, mock_global_config):
@@ -155,7 +155,7 @@ class TestSecretsAddCLIForceFlag:
         # Verify multiple repos were passed
         assert len(call_kwargs.get('repo_urls', [])) == 3
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     @patch('classdock.services.secrets_service.get_global_config')
     @patch('classdock.secrets.github_secrets.GitHubSecretsManager')
     def test_secrets_add_force_without_repos(self, mock_manager_class, mock_get_config_service, mock_get_config_cli, runner, mock_global_config):
@@ -201,7 +201,7 @@ class TestSecretsAddCLIForceFlag:
         assert 'force' in output_lower and 'update' in output_lower, \
             f"Expected force update description in output, but got: {clean_output[:500]}"
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     def test_secrets_add_force_with_no_config(self, mock_get_config, runner):
         """Test --force flag behavior when no config is available."""
         mock_get_config.return_value = None
@@ -216,7 +216,7 @@ class TestSecretsAddCLIForceFlag:
         assert 'not loaded' in result.output.lower(
         ) or 'configuration' in result.output.lower()
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     def test_secrets_add_force_with_no_secrets_config(self, mock_get_config, runner):
         """Test --force flag when no secrets are configured."""
         mock_config = MagicMock(spec=GlobalConfig)
@@ -236,7 +236,7 @@ class TestSecretsAddCLIForceFlag:
 class TestSecretsAddCLIIntegration:
     """Integration tests for secrets add CLI with force flag."""
 
-    @patch('classdock.cli.get_global_config')
+    @patch('classdock.commands.secrets.get_global_config')
     @patch('classdock.services.secrets_service.get_global_config')
     @patch('classdock.secrets.github_secrets.GitHubSecretsManager')
     def test_complete_workflow_with_force(self, mock_manager_class, mock_get_config_service, mock_get_config_cli, runner, mock_global_config):
